Move (nexus-core) org.sonatype.nexus.util.sequence.* to (nexus-common) org.sonatype.nexus.common.sequence.* Note we may remove this in the future, but for now there are a few places using this, so to help clean up nexus-core, move to common
