+++ components/nexus-common/src/test/java/org/sonatype/nexus/common/sequence/NumberSequenceTest.java
@@ -10,12 +10,18 @@
  * of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
  * Eclipse Foundation. All other trademarks are the property of their respective owners.
  */
-package org.sonatype.nexus.util.sequence;
+package org.sonatype.nexus.common.sequence;
+
+import org.sonatype.sisu.litmus.testsupport.TestSupport;
 
 import org.junit.Assert;
 import org.junit.Test;
 
+/**
+ * Tests for various {@link NumberSequence} implementations.
+ */
 public class NumberSequenceTest
+  extends TestSupport
 {
   @Test
   public void testConstantSequence() {
