+++ components/nexus-core/src/main/java/org/sonatype/nexus/proxy/repository/AbstractProxyRepository.java
@@ -25,6 +25,9 @@ import javax.net.ssl.SSLException;
 import javax.net.ssl.SSLPeerUnverifiedException;
 
 import org.sonatype.nexus.common.property.SystemPropertiesHelper;
+import org.sonatype.nexus.common.sequence.ConstantNumberSequence;
+import org.sonatype.nexus.common.sequence.FibonacciNumberSequence;
+import org.sonatype.nexus.common.sequence.NumberSequence;
 import org.sonatype.nexus.configuration.model.CRemoteStorage;
 import org.sonatype.nexus.proxy.IllegalOperationException;
 import org.sonatype.nexus.proxy.ItemNotFoundException;
@@ -62,9 +65,6 @@ import org.sonatype.nexus.proxy.utils.RepositoryStringUtils;
 import org.sonatype.nexus.proxy.walker.DefaultWalkerContext;
 import org.sonatype.nexus.proxy.walker.WalkerException;
 import org.sonatype.nexus.proxy.walker.WalkerFilter;
-import org.sonatype.nexus.util.sequence.ConstantNumberSequence;
-import org.sonatype.nexus.util.sequence.FibonacciNumberSequence;
-import org.sonatype.nexus.util.sequence.NumberSequence;
 
 import com.google.common.base.Throwables;
 import com.google.common.collect.Lists;
